+++ app/services/RabbitmqPlugin.scala
@@ -20,6 +20,7 @@ import play.api.libs.json._
 import play.api.libs.ws.{Response, WS}
 import play.api.{Application, Logger, Plugin}
 import play.libs.Akka
+import securesocial.core.IdentityId
 
 import scala.concurrent.{Await, Future}
 import scala.util.Try
@@ -83,6 +84,7 @@ class RabbitmqPlugin(application: Application) extends Plugin {
   val spacesService: SpaceService = DI.injector.getInstance(classOf[SpaceService])
   val extractorsService: ExtractorService = DI.injector.getInstance(classOf[ExtractorService])
   val datasetService: DatasetService = DI.injector.getInstance(classOf[DatasetService])
+  val userService: UserService = DI.injector.getInstance(classOf[UserService])
 
   var extractQueue: Option[ActorRef] = None
   val cancellationDownloadQueueName: String = "clowder.jobs.temp"
@@ -100,7 +102,7 @@ class RabbitmqPlugin(application: Application) extends Plugin {
   var mgmtPort: String = ""
   var bindings = List.empty[Binding]
 
-  var apiKey = play.api.Play.configuration.getString("commKey").getOrElse("")
+  var globalAPIKey = play.api.Play.configuration.getString("commKey").getOrElse("")
 
   override def onStart() {
     Logger.info("Starting RabbitMQ Plugin")
@@ -402,15 +404,34 @@ class RabbitmqPlugin(application: Application) extends Plugin {
     }
   }
 
+
+  /** Return the API key to use in the submission. If the one in the key is not set in the request then get the default
+    * extraction key for the user. If the user is not defined default to the global key.
+    * for the user
+    *
+    * @param requestAPIKey the API key from the request
+    * @param user the user from the request
+    * @return the API key to use
+    */
+  def getApiKey(requestAPIKey: Option[String], user: Option[User]): String = {
+    if (requestAPIKey.isDefined)
+      requestAPIKey.get
+    else if (user.isDefined)
+      userService.getExtractionApiKey(user.get.identityId).key
+    else
+      globalAPIKey
+  }
+
   /**
     * Publish to the proper queues when a new file is uploaded to the system.
     * @param file the file that was just uploaded
     * @param dataset the dataset the file belongs to
     * @param host the Clowder host URL for sharing extractors across instances
     */
-  def fileCreated(file: File, dataset: Option[Dataset], host: String): Unit = {
+  def fileCreated(file: File, dataset: Option[Dataset], host: String, requestAPIKey: Option[String]): Unit = {
     val routingKey = exchange + "." + "file." + contentTypeToRoutingKey(file.contentType)
     val extraInfo = Map("filename" -> file.filename)
+    val apiKey = requestAPIKey.getOrElse(globalAPIKey)
     val sourceExtra = JsObject((Seq("filename" -> JsString(file.filename))))
     Logger.debug(s"Sending message $routingKey from $host with extraInfo $extraInfo")
     dataset match {
@@ -440,9 +461,10 @@ class RabbitmqPlugin(application: Application) extends Plugin {
     * @param file the file that was just uploaded
     * @param host the Clowder host URL for sharing extractors across instances
     */
-  def fileCreated(file: TempFile, host: String): Unit = {
+  def fileCreated(file: TempFile, host: String, requestAPIKey: Option[String]): Unit = {
     val routingKey = exchange + "." + "file." + contentTypeToRoutingKey(file.contentType)
     val extraInfo = Map("filename" -> file.filename)
+    val apiKey = requestAPIKey.getOrElse(globalAPIKey)
     Logger.debug(s"Sending message $routingKey from $host with extraInfo $extraInfo")
     val sourceExtra = JsObject((Seq("filename" -> JsString(file.filename))))
     val source = Entity(ResourceRef(ResourceRef.file, file.id), Some(file.contentType), sourceExtra)
@@ -461,10 +483,11 @@ class RabbitmqPlugin(application: Application) extends Plugin {
     * @param dataset the dataset it was added to
     * @param host the Clowder host URL for sharing extractors across instances
     */
-  def fileAddedToDataset(file: File, dataset: Dataset, host: String): Unit = {
+  def fileAddedToDataset(file: File, dataset: Dataset, host: String, requestAPIKey: Option[String]): Unit = {
     val routingKey = s"$exchange.dataset.file.added"
     Logger.debug(s"Sending message $routingKey from $host")
     val queues = getQueues(dataset, routingKey, file.contentType)
+    val apiKey = requestAPIKey.getOrElse(globalAPIKey)
     val sourceExtra = JsObject((Seq("filename" -> JsString(file.filename))))
     queues.foreach{ extractorId =>
       val source = Entity(ResourceRef(ResourceRef.file, file.id), Some(file.contentType), sourceExtra)
@@ -490,10 +513,11 @@ class RabbitmqPlugin(application: Application) extends Plugin {
     * @param dataset the dataset it was added to
     * @param host the Clowder host URL for sharing extractors across instances
     */
-  def fileSetAddedToDataset(dataset: Dataset, filelist: List[File], host: String): Unit = {
+  def fileSetAddedToDataset(dataset: Dataset, filelist: List[File], host: String, requestAPIKey: Option[String]): Unit = {
     val routingKey = s"$exchange.dataset.files.added"
     Logger.debug(s"Sending message $routingKey from $host")
     val queues = getQueues(dataset, routingKey, "")
+    val apiKey = requestAPIKey.getOrElse(globalAPIKey)
     val sourceExtra = JsObject((Seq("filenames" -> JsArray(filelist.map(f=>JsString(f.filename)).toSeq))))
     val msgExtra = Map("filenames" -> filelist.map(f=>f.filename))
     queues.foreach{ extractorId =>
@@ -520,10 +544,11 @@ class RabbitmqPlugin(application: Application) extends Plugin {
     * @param dataset
     * @param host
     */
-  def fileRemovedFromDataset(file: File, dataset: Dataset, host: String): Unit = {
+  def fileRemovedFromDataset(file: File, dataset: Dataset, host: String, requestAPIKey: Option[String]): Unit = {
     val routingKey = s"$exchange.dataset.file.removed"
     Logger.debug(s"Sending message $routingKey from $host")
     val queues = getQueues(dataset, routingKey, file.contentType)
+    val apiKey = requestAPIKey.getOrElse(globalAPIKey)
     val sourceExtra = JsObject((Seq("filename" -> JsString(file.filename))))
     queues.foreach{ extractorId =>
       val source = Entity(ResourceRef(ResourceRef.file, file.id), Some(file.contentType), sourceExtra)
@@ -551,8 +576,9 @@ class RabbitmqPlugin(application: Application) extends Plugin {
     * @param newFlags
     */
   def submitFileManually(originalId: UUID, file: File, host: String, queue: String, extraInfo: Map[String, Any],
-                         datasetId: UUID, newFlags: String): Unit = {
+    datasetId: UUID, newFlags: String, requestAPIKey: Option[String], user: Option[User]): Unit = {
     Logger.debug(s"Sending message to $queue from $host with extraInfo $extraInfo")
+    val apiKey = getApiKey(requestAPIKey, user)
     val sourceExtra = JsObject((Seq("filename" -> JsString(file.filename))))
     val source = Entity(ResourceRef(ResourceRef.file, originalId), Some(file.contentType), sourceExtra)
 
@@ -574,8 +600,10 @@ class RabbitmqPlugin(application: Application) extends Plugin {
     * @param datasetId
     * @param newFlags
     */
-  def submitDatasetManually(host: String, queue: String, extraInfo: Map[String, Any], datasetId: UUID, newFlags: String, user: Option[User]): Unit = {
+  def submitDatasetManually(host: String, queue: String, extraInfo: Map[String, Any], datasetId: UUID, newFlags: String,
+    requestAPIKey: Option[String], user: Option[User]): Unit = {
     Logger.debug(s"Sending message $queue from $host with extraInfo $extraInfo")
+    val apiKey = getApiKey(requestAPIKey, user)
     val source = Entity(ResourceRef(ResourceRef.dataset, datasetId), None, JsObject(Seq.empty))
 
     val notifies = user match {
@@ -598,9 +626,12 @@ class RabbitmqPlugin(application: Application) extends Plugin {
     * @param extraInfo
     * @param host
     */
-  def metadataAddedToResource(metadataId: UUID, resourceRef: ResourceRef, extraInfo: Map[String, Any], host: String): Unit = {
+  // FIXME check if extractor is enabled in space or global
+  def metadataAddedToResource(metadataId: UUID, resourceRef: ResourceRef, extraInfo: Map[String, Any], host: String,
+    requestAPIKey: Option[String], user: Option[User]): Unit = {
     val routingKey = s"$exchange.metadata.added.${resourceRef.resourceType.name}"
     Logger.debug(s"Sending message $routingKey from $host with extraInfo $extraInfo")
+    val apiKey = getApiKey(requestAPIKey, user)
 
     resourceRef.resourceType match {
       // metadata added to dataset
@@ -648,8 +679,9 @@ class RabbitmqPlugin(application: Application) extends Plugin {
     * @param host
     */
   // FIXME check if extractor is enabled in space or global
-  def metadataRemovedFromResource(metadataId: UUID, resourceRef: ResourceRef, host: String): Unit = {
+  def metadataRemovedFromResource(metadataId: UUID, resourceRef: ResourceRef, host: String, requestAPIKey: Option[String], user: Option[User]): Unit = {
     val routingKey = s"$exchange.metadata.removed.${resourceRef.resourceType.name}"
+    val apiKey = getApiKey(requestAPIKey, user)
     val extraInfo = Map[String, Any](
       "resourceType"->resourceRef.resourceType.name,
       "resourceId"->resourceRef.id.toString)
@@ -701,10 +733,11 @@ class RabbitmqPlugin(application: Application) extends Plugin {
     * @param length
     * @param host
     */
-  def multimediaQuery(tempFileId: UUID, contentType: String, length: String, host: String): Unit = {
+  def multimediaQuery(tempFileId: UUID, contentType: String, length: String, host: String, requestAPIKey: Option[String]): Unit = {
     //key needs to contain 'query' when uploading a query
     //since the thumbnail extractor during processing will need to upload to correct mongo collection.
     val routingKey = exchange +".query." + contentType.replace("/", ".")
+    val apiKey = requestAPIKey.getOrElse(globalAPIKey)
     Logger.debug(s"Sending message $routingKey from $host")
     //FIXME, should we pass user's email address?
     val notifies = List[String]()
@@ -722,8 +755,9 @@ class RabbitmqPlugin(application: Application) extends Plugin {
     * @param sectionId
     * @param host
     */
-  def submitSectionPreviewManually(preview: Preview, sectionId: UUID, host: String): Unit = {
+  def submitSectionPreviewManually(preview: Preview, sectionId: UUID, host: String, requestAPIKey: Option[String]): Unit = {
     val routingKey = exchange + ".index."+ contentTypeToRoutingKey(preview.contentType)
+    val apiKey = requestAPIKey.getOrElse(globalAPIKey)
     val extraInfo = Map("section_id"->sectionId)
     val source = Entity(ResourceRef(ResourceRef.preview, preview.id), None, JsObject(Seq.empty))
     val target = Entity(ResourceRef(ResourceRef.section, sectionId), None, JsObject(Seq.empty))