+++ app/services/mongodb/MongoDBSpaceService.scala
@@ -300,11 +300,11 @@ class MongoDBSpaceService @Inject() (
     ProjectSpaceDAO.save(space)
   }
 
-  def delete(id: UUID, host: String): Unit = {
+  def delete(id: UUID, host: String, apiKey: Option[String], user: Option[User]): Unit = {
     // only curation objects in this space are removed, since dataset & collection don't need to belong to a space.
     get(id) match {
       case Some(s) => {
-        s.curationObjects.map(c => curations.remove(c, host))
+        s.curationObjects.map(c => curations.remove(c, host, apiKey, user))
         for(follower <- s.followers) {
           users.unfollowResource(follower, ResourceRef(ResourceRef.space, id))
         }
@@ -446,7 +446,7 @@ class MongoDBSpaceService @Inject() (
    *  @param space The id of the space to check
    *
    */
-  def purgeExpiredResources(space: UUID, host: String): Unit = {
+  def purgeExpiredResources(space: UUID, host: String, apiKey: Option[String], user: Option[User]): Unit = {
       val datasetsList = getDatasetsInSpace(Some(space.stringify))
       val collectionsList = getCollectionsInSpace(Some(space.stringify))
       val timeToLive = getTimeToLive(space)
@@ -457,7 +457,7 @@ class MongoDBSpaceService @Inject() (
     	  val difference = currentTime - datasetTime
     	  if (difference > timeToLive) {
     	       //It was last modified longer than the time to live, so remove it.
-    	       datasets.removeDataset(aDataset.id, host)
+    	       datasets.removeDataset(aDataset.id, host, apiKey, user)
     	  }
       }
 
@@ -486,7 +486,7 @@ class MongoDBSpaceService @Inject() (
                 datasetOnlyInSpace match {
                   case Some(true) => {
                     //If the dataset only exists in the current space, it can be removed.
-                    datasets.removeDataset(colDataset.id, host)
+                    datasets.removeDataset(colDataset.id, host, apiKey, user)
                   }
                   case None => {
                     //In this case, the dataset is in the default space, so do not remove it, it will detach on collection deletion.